17 задача

+++ artifacts/models.py
@@ -7,7 +7,7 @@ class CollectibleItem(models.Model):
     """Модель для артефактов которые можно собрать в забеге"""
     name = models.CharField(max_length=200)
     uid = models.UUIDField(default=uuid.uuid4, editable=False, unique=True)
-    latitude = models.DecimalField()
-    longitude = models.DecimalField()
+    latitude = models.FloatField()
+    longitude = models.FloatField()
     picture = models.URLField()
     value = models.IntegerField()
